# GardenControlUI/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GardenControlApi",
    "description": "An API that allows the monitor and control of devices plugged in to the GPIO Pins on a Raspberry Pi",
    "version": "v1"
  },
  "paths": {
...
    "/api/ControlDevice": {
      "get": {
        "tags": [
          "ControlDevice"
        ],
        "summary": "Returns all Control Devices",
        "responses": {
          "200": {
            "description": "Returns all control devices",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ControlDeviceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ControlDeviceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ControlDeviceDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ControlDevice"
        ],
        "summary": "Creates a Control Device",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Control Device created successfully"
          }
        }
      }
    },
    "/api/ControlDevice/{id}": {
      "get": {
        "tags": [
          "ControlDevice"
        ],
        "summary": "Returns a Control Device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the specified control device",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ControlDeviceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlDeviceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlDeviceDto"
                }
              }
            }
          },
          "404": {
            "description": "Could not find a control device with the specified id",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ControlDevice"
        ],
        "summary": "Updates a Control Device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ControlDeviceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Control Device successfully updated"
          },
          "400": {
            "description": "Control Device Id in url and object do not match",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Could not find a control device with the specified id",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ControlDevice"
        ],
        "summary": "Deletes a Control Device",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Control Device successfully deleted"
          },
          "404": {
            "description": "Could not find a control device with the specified id",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/DS18B20/{id}": {
      "get": {
        "tags": [
          "DS18B20"
        ],
        "summary": "Returns a reading value from the specified DS18B20 temperature probe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns temperature value",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DS18B20Dto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DS18B20Dto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DS18B20Dto"
                }
              }
            }
          },
          "404": {
            "description": "Could not find DS18B20 probe from id",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/DS18B20/SerialNumbers/{gpioPin}": {
      "get": {
        "tags": [
          "DS18B20"
        ],
        "summary": "Returns the serial numbers of all the attached probes to the specified GPIO Pin.\r\nUseful for getting serial numbers to register each probe.",
        "parameters": [
          {
            "name": "gpioPin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of the 1-wire protocol serial numbers attached to the GPIO pin",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "DeviceType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ControlDeviceDto": {
        "type": "object",
        "properties": {
          "controlDeviceId": {
            "type": "integer",
            "format": "int32"
          },
          "deviceTypeId": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "deviceType": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "alias": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "gpioPinNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "serialNumber": {
            "type": "string",
            "nullable": true
          },
          "defaultState": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DS18B20Dto": {
        "type": "object",
        "properties": {
          "readingDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "readingValueCelcius": {
            "type": "number",
            "format": "double"
          },
          "readingValueFahrenheit": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}